+++ CHANGELOG.md
@@ -25,3 +25,4 @@
 - Implement `xsd:unsignedLong` data type codec
 - Implement `xsd:gDay` codec and data type
 - Implement `xsd:gMonthDay` codec and data type
+- Implement `xsd:duration` codec and data type